circuitos desde coap

+++ VeltiumCharger/main/group_control.cpp
@@ -297,8 +297,8 @@ void New_Group(char* Data, int Data_size){
 
         add_to_group(ID, get_IP(ID), charger_table, &ChargingGroup.Charger_number);
 
-        charger_table[i].Fase = (Data[10+i*9]-'0')& 0x03;
-        charger_table[i].Circuito = (Data[10+i*9]-'0') >> 0x02;
+        charger_table[i].Fase = uint8_t(Data[10+i*9]-'0') & 0x03;
+        charger_table[i].Circuito = uint8_t(Data[10+i*9]-'0') >> 0x02;
 
         uint8_t index =check_in_group(ID, temp_chargers, temp_chargers_size);
         if(index != 255){
@@ -314,7 +314,7 @@ void New_Group(char* Data, int Data_size){
 
         if(!memcmp(ID,ConfigFirebase.Device_Id,8)){
             charger_table[i].Conected = true;
-            Params.Fase = (Data[10+i*9]-'0')& 0x03;
+            Params.Fase = uint8_t(Data[10+i*9]-'0') & 0x03;
         }
     }
